added fix

+++ modules/db.py
@@ -70,7 +70,8 @@ class EnhancedPicklePersistence(PicklePersistence):
             self._load_chats_from_json()
             self.bot_data = self.context_types.bot_data()
             self.callback_data = None
-            os.rename('./db/db.json', './db/db.json.bak')
+            if os.path.exists('./db/db.json'):
+                os.rename('./db/db.json', './db/db.json.bak')
             self._dump_singlefile()
         except pickle.UnpicklingError as exc:
             filename = self.filepath.name
